Fix: Usersテーブルの主キー型を修正

+++ database/migrations/2020_09_28_104152_change_users_table.php
@@ -14,7 +14,7 @@ class ChangeUsersTable extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->bigInteger('id')->change();
+            $table->bigIncrements('id')->change();
         });
     }
 
@@ -26,7 +26,8 @@ class ChangeUsersTable extends Migration
     public function down()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->int('id')->change();
+            // $table->int('id')->change();
+            $table->bigInteger('id')->change();
         });
     }
 }
